requirements.txt
google-generativeai>=0.7.0
streamlit>=1.32.0
pillow>=10.3.0
python-dotenv>=1.0.0
requests>=2.31.0